refactor Panels and NodePanels
