+++ components/interview/interfaces/name-generator/NodePanel.tsx
@@ -1,16 +1,52 @@
-import Panel from './Panel';
-import NodeList, { type Node } from '../../NodeList';
+import * as Accordion from '@radix-ui/react-accordion';
+import { motion } from 'framer-motion';
+import { MotionSurface } from '~/components/layout/Surface';
+import Heading from '~/components/typography/Heading';
+import { cn } from '~/lib/utils';
 
-export default function NodePanel({
-  nodes,
+// incompatibility between framer-motion 12.x and new react types
+// eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+const MotionTrigger = motion(Accordion.Trigger);
+
+export default function Panel({
+  id,
   title,
+  expanded,
+  children,
+  noHighlight,
 }: {
-  nodes: Node[];
+  id: string;
   title: string;
+  expanded: boolean;
+  children: React.ReactNode;
+  noHighlight?: boolean;
 }) {
+  const panelClasses = cn(
+    'flex flex-1 flex-col rounded-small border-b border-b-4 border-panel-1 shadow-xl',
+    {
+      'border-b-0': noHighlight,
+    },
+  );
+
+  const contentClasses = cn(
+    'h-auto flex-grow border-t border-background flex flex-col overflow-hidden items-center',
+  );
+
   return (
-    <Panel title={title}>
-      <NodeList items={nodes} nodeSize="sm" />
-    </Panel>
+    <Accordion.Item value={id} asChild>
+      <MotionSurface layout level={1} spacing="none" className={panelClasses}>
+        <MotionTrigger
+          layout
+          className="flex cursor-pointer items-center justify-center p-4"
+        >
+          <Heading variant="h3" className="mb-0">
+            {title}
+          </Heading>
+        </MotionTrigger>
+        <Accordion.Content className={contentClasses}>
+          {expanded ? children : null}
+        </Accordion.Content>
+      </MotionSurface>
+    </Accordion.Item>
   );
 }
